requirements.txt
Django==2.2.3
imageio==2.5.0
numpy==1.16.4
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
